requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Data processing
pandas==2.1.4
numpy==1.26.3

# Location services
geopy==2.4.1

# AI/OpenAI
openai==1.9.0

# HTTP client
httpx==0.26.0

# Environment and config
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4